pyproject.toml
[tool.poetry]
name = "fast-api-strawberry-docker"
version = "0.1.0"
description = ""
authors = ["FabianGoessling <Fabian.Goessling@me.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.75.1"
strawberry-graphql = "^0.106.3"
uvicorn = {extras = ["standard"], version = "^0.17.6"}

[tool.poetry.dev-dependencies]
black = "^22.3.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
